fix: change view at parent component
